+++ src/SMAPI/Framework/SGame.cs
@@ -12,7 +12,6 @@ using Microsoft.Xna.Framework.Input;
 using StardewModdingAPI.Events;
 using StardewModdingAPI.Framework.Reflection;
 using StardewModdingAPI.Framework.Utilities;
-using StardewModdingAPI.Utilities;
 using StardewValley;
 using StardewValley.BellsAndWhistles;
 using StardewValley.Locations;
@@ -371,7 +370,8 @@ namespace StardewModdingAPI.Framework
                     SButton[] previousPressedKeys = this.PreviousPressedButtons;
                     SButton[] framePressedKeys = currentlyPressedKeys.Except(previousPressedKeys).ToArray();
                     SButton[] frameReleasedKeys = previousPressedKeys.Except(currentlyPressedKeys).ToArray();
-                    bool isClick = framePressedKeys.Contains(SButton.MouseLeft) || (framePressedKeys.Contains(SButton.ControllerA) && !currentlyPressedKeys.Contains(SButton.ControllerX));
+                    bool isUseToolButton = Game1.options.useToolButton.Any(p => framePressedKeys.Contains(p.ToSButton()));
+                    bool isActionButton = !isUseToolButton && Game1.options.actionButton.Any(p => framePressedKeys.Contains(p.ToSButton()));
 
                     // get cursor position
                     ICursorPosition cursor;
@@ -388,7 +388,7 @@ namespace StardewModdingAPI.Framework
                     // raise button pressed
                     foreach (SButton button in framePressedKeys)
                     {
-                        InputEvents.InvokeButtonPressed(this.Monitor, button, cursor, isClick);
+                        InputEvents.InvokeButtonPressed(this.Monitor, button, cursor, isActionButton, isUseToolButton);
 
                         // legacy events
                         if (button.TryGetKeyboard(out Keys key))
@@ -408,10 +408,9 @@ namespace StardewModdingAPI.Framework
                     // raise button released
                     foreach (SButton button in frameReleasedKeys)
                     {
-                        bool wasClick =
-                            (button == SButton.MouseLeft && previousPressedKeys.Contains(SButton.MouseLeft)) // released left click
-                            || (button == SButton.ControllerA && previousPressedKeys.Contains(SButton.ControllerA) && !previousPressedKeys.Contains(SButton.ControllerX));
-                        InputEvents.InvokeButtonReleased(this.Monitor, button, cursor, wasClick);
+                        bool wasUseToolButton = (from opt in Game1.options.useToolButton let optButton = opt.ToSButton() where optButton == button && framePressedKeys.Contains(optButton) select optButton).Any();
+                        bool wasActionButton = !wasUseToolButton && (from opt in Game1.options.actionButton let optButton = opt.ToSButton() where optButton == button && framePressedKeys.Contains(optButton) select optButton).Any();
+                        InputEvents.InvokeButtonReleased(this.Monitor, button, cursor, wasActionButton, wasUseToolButton);
 
                         // legacy events
                         if (button.TryGetKeyboard(out Keys key))